+++ app/project/bootstrappers/http/views/BuildersBootstrapper.php
@@ -8,6 +8,7 @@ use Project\HTTP\Views\Builders\HomeBuilder;
 use Project\HTTP\Views\Builders\MasterBuilder;
 use Opulence\Applications\Bootstrappers\Bootstrapper;
 use Opulence\Views\Factories\IViewFactory;
+use Opulence\Views\IView;
 
 class BuildersBootstrapper extends Bootstrapper
 {
@@ -18,11 +19,11 @@ class BuildersBootstrapper extends Bootstrapper
      */
     public function run(IViewFactory $viewFactory)
     {
-        $viewFactory->registerBuilder("Master", function () {
-            return new MasterBuilder();
+        $viewFactory->registerBuilder("Master", function (IView $view) {
+            return (new MasterBuilder())->build($view);
         });
-        $viewFactory->registerBuilder("Home", function () {
-            return new HomeBuilder();
+        $viewFactory->registerBuilder("Home", function (IView $view) {
+            return (new HomeBuilder())->build($view);
         });
     }
 }